==> 07-Yelp-Rating-Reco.ipynb <==
### 07-Yelp-Reco.ipynb
* **Script**: 07-Yelp-Rating-Reco.ipynb
* **Purpose**: Multicriteria Recommendations with ratings
* **Input**: data/yelp/dataframes/yelp_review_user.pkl
*            data/yelp/dataframes/yelp_review_business.pkl
*            data/yelp/dataframes/yelp_review_user_business.pkl
* **Output**: reco_with_rating_output.txt

import warnings
import numpy as np
import pandas as pd
import seaborn as sns
from matplotlib import pyplot as plt
%matplotlib inline

from sklearn.decomposition import PCA
from sklearn.cluster import MiniBatchKMeans, DBSCAN

warnings.filterwarnings("ignore", category=DeprecationWarning)

#Function to Print Percentiles (for Cleaning)
def percentiles(df, variable):
    print("Min:  "+str(df[variable].quantile(q=0.00)))
    print("1st:  "+str(df[variable].quantile(q=0.01)))
    print("5th:  "+str(df[variable].quantile(q=0.05)))
    print("10th: "+str(df[variable].quantile(q=0.10)))
    print("25th: "+str(df[variable].quantile(q=0.25)))
    print("50th: "+str(df[variable].quantile(q=0.50)))
    print("75th: "+str(df[variable].quantile(q=0.75)))
    print("90th: "+str(df[variable].quantile(q=0.90)))
    print("95th: "+str(df[variable].quantile(q=0.95)))
    print("99th: "+str(df[variable].quantile(q=0.99)))
    print("Max:  "+str(df[variable].quantile(q=1.00)))

def getbounds_75_percentiles(df, variable):
    return df[variable].quantile(q=0.75), df[variable].quantile(q=0.99)

User Statistics

user = pd.read_pickle("data/yelp/dataframes/yelp_review_user.pkl")
user.head(5)

print("Number of users", user["user_id"].nunique())

Business Statistics

business = pd.read_pickle("data/yelp/dataframes/yelp_review_business.pkl")
business.head(5)

print("Number of businesses", business["business_id"].nunique())

user_business = pd.read_pickle("data/yelp/dataframes/yelp_review_user_business.pkl")
user_business.head(5)

a, b = getbounds_75_percentiles(user, "relevant")
print("Bounds for 75 to 99 percentils are:", a, b)
user = user[(user.relevant>=a) & (user.relevant<=b)]
user.reset_index(drop=True, inplace=True)

user["pct_food_positive"] = user["topic_food_positive"] / user["relevant"]
user["pct_food_negative"] = user["topic_food_negative"] / user["relevant"]
user["pct_service_positive"] = user["topic_service_positive"] / user["relevant"]
user["pct_service_negative"] = user["topic_service_negative"] / user["relevant"]
user["pct_ambience_positive"] = user["topic_ambience_positive"] / user["relevant"]
user["pct_ambience_negative"] = user["topic_ambience_negative"] / user["relevant"]
user["pct_value_positive"] = user["topic_value_positive"] / user["relevant"]
user["pct_value_negative"] = user["topic_value_negative"] / user["relevant"]

user.head(5)

user_aspects = pd.DataFrame(user, columns = ["user_id", "pct_food_positive", "pct_food_negative", "pct_service_positive", "pct_service_negative", "pct_ambience_positive", "pct_ambience_negative", "pct_value_positive","pct_value_negative"])
user_aspects.head(5)

print("Size of user aspects vector",  user_aspects.size)

user_vector = user_aspects.groupby(['user_id']).agg(['mean'])
uv = user_vector.head(5)
uv

c, d = getbounds_75_percentiles(business, "relevant")
print("Bounds for 75 to 99 percentils are:", c, d)
business = business[(business.relevant>=c) & (business.relevant<=d)]
business.reset_index(drop=True, inplace=True)

business["pct_food_positive"] = business["topic_food_positive"] / business["relevant"]
business["pct_food_negative"] = business["topic_food_negative"] / business["relevant"]
business["pct_service_positive"] = business["topic_service_positive"] / business["relevant"]
business["pct_service_negative"] = business["topic_service_negative"] / business["relevant"]
business["pct_ambience_positive"] = business["topic_ambience_positive"] / business["relevant"]
business["pct_ambience_negative"] = business["topic_ambience_negative"] / business["relevant"]
business["pct_value_positive"] = business["topic_value_positive"] / business["relevant"]
business["pct_value_negative"] = business["topic_value_negative"] / business["relevant"]

business.head(5)

business_aspects = pd.DataFrame(business, columns = ["business_id", "pct_food_positive", "pct_food_negative", "pct_service_positive", "pct_service_negative", "pct_ambience_positive", "pct_ambience_negative", "pct_value_positive", "pct_value_negative"])
business_aspects.head(5)

print("Size of business aspects vector",  business_aspects.size)

business_vector = business_aspects.groupby(['business_id']).agg(['mean'])
bv = business_vector.head(5)
bv

#### Generate recommendations with user_aspects vector and business_aspects vector

from scipy import spatial

no_Of_Users = 25
count = 0

def distance(x,y):
    return (spatial.distance.cosine(x[1::2],y[1:9:2]) + (1 - spatial.distance.cosine(x[2::2],y[2:9:2])))
    
f = open('/home/nilanjan/Projects/MotifPolitik/PoC/Code/ABSA/reco_with_rating_output.txt', 'w')
    
for u in user_vector.itertuples():
    if (count >= no_Of_Users):
        break
    else:
        #val = "User id: " + str(u[0])
        f.write("\n\n")
        f.write("User Aspect Vector: ")
        f.write("\n\n")
        f.write(str(u))
        f.write("\n\n")
        for b in business_vector.itertuples():
            business_vector.loc[b[0],'score'] = distance(u,b)
            uratings = user_business['rating'][(user_business['user_id'] == str(u[0]))&(user_business['business_id'] == str(b[0]))].values
            if (uratings.size > 0):
                rarray = uratings[0]
                if (rarray.size > 0):
                    rating = rarray[0]
                else: rating = 0
            else: rating = 0
            business_vector.loc[b[0],'rating'] = rating
            #print("Size of business vector: " + str(business_vector.shape))
            #bv_ratings = business_vector['rating'] > 0
            #print("Size of bv ratings: " + str(bv_ratings.shape))
        f.write(business_vector.ix[business_vector.rating > 0,:].sort_values(by="score").to_string())
        count += 1
f.close()

print("Recommendations printed in reco_with_rating_output.txt")
